refactor: rename PanelLeft components
