+++ src/components/PanelLeft/PaneDatabase.vue
